# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '0.1'
  title: 'Gamma Public API'
  description: 'Web Resources Specification (A7) for Gamma'
servers:
  - url: https://lbaw2391.lbaw.fe.up.pt
    description: Production server
...
paths:
...
  /api/admin/search/users/{query}:

    get:
      operationId: R305
      summary: 'R305: Previews the users search result to use for AJAX in the admin page.' 
      description: 'Allows the search for users. Access: ADM'
      parameters:
        - in: path
          name: query
          description: String to use for full-text search
          schema:
            type: string
          required: false
      tags:
        - 'M03: User Administration'
      
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    image:
                      type: string
                    username:
                      type: string
                    e-mail:
                      type: string
                    academic_status:
                      type: string
                    display_name:
                      type: string
                    is_private:
                      type: boolean
                example:
                  - id: 1
                    image: url/of/the/image.jpg
                    username: john_doe
                    e-mail: johndoe@trigger.com
                    academic_status: Student 
                    display_name: John Doe
                    is_private: False
                  - id: 28
                    image: url/of/the/image2.jpg
                    username: carlosbarbossa
                    e-mail: carlosbarbossa@carlosbarbossa.com
                    academic_status: Student 
                    display_name: Carlos Barbossa
                    is_private: False
  /users/{username}:

    get:
      operationId: R501
      summary: 'R501: Profile Action.'
      description: 'Get the page view for a user profile. Access: VST, USR'
      tags:
        - 'M05: Users'
      responses:
        '200':
          description: 'OK. Show User Profile'
        '403':
          description: 'Forbidden. User doesnt have access to this profile.'

      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          
    delete:
      operationId: R502
      summary: 'R502: Deletes an user'
      description: 'Allows an authenticated user or admin to delete himself or another user (if admin). Access: USR, ADM'
      tags:
        - 'M05: Users'

      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
    
      responses:
        '200':
          description: Success
        '422':
          description: Failure. If user not found with username
        '403':
          description: Action forbidden
...
  /api/search/users/{query}:

    get:
      operationId: R602
      summary: 'R602: Previews the users search result to use for AJAX'
      description: 'Allows the search for users. Access: VST, USR'
      tags:
        - 'M06: Search'
      parameters:
      - in: path
        name: query
        description: String to use for full-text search
        schema:
          type: string
        required: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    image:
                      type: string
                    username:
                      type: string
                    e-mail:
                      type: string
                    academic_status:
                      type: string
                    display_name:
                      type: string
                    is_private:
                      type: boolean
                example:
                  - id: 1
                    image: url/of/the/image.jpg
                    username: john_doe
                    e-mail: johndoe@trigger.com
                    academic_status: Student 
                    display_name: John Doe
                    is_private: False
                  - id: 28
                    image: url/of/the/image2.jpg
                    username: carlosbarbossa
                    e-mail: carlosbarbossa@carlosbarbossa.com
                    academic_status: Student 
                    display_name: Carlos Barbossa
                    is_private: False
...